[cleanup] Reorganizes utils into database and ping  - dbUtils in development now affects production server
