+++ server/api/apiController.js
@@ -1,13 +1,13 @@
 var testData = require('../data/testData');
-var dbUtils = require('../utils/dbUtils');
+var dbUtils = require('../utils/dbUtils/dbUtils');
 
 // Set to true to use the database or false to use test data
-var useDB = true;
+var useDb = true;
 
-// These functions have else branches in the case of useDB false
+// These functions have else branches in the case of useDb false
 module.exports = {
   getDashboardInfo: function(req, res, next) {
-    if(useDB) {
+    if(useDb) {
       var user = {id: req.query.id};
       dbUtils.getUsersShareOrganization(user, function(error, results) {
         res.end(JSON.stringify(results));
@@ -20,7 +20,7 @@ module.exports = {
   },
 
   getClientInfo: function(req, res, next) {
-    if(useDB) {
+    if(useDb) {
       var user = {id: req.query.id};
       dbUtils.getUsersShareOrganization(user, function(error, results) {
         res.end(JSON.stringify(results));
