+++ server/config/config.example.js
@@ -4,10 +4,20 @@ module.exports = {
     username: 'YourNameHere',
     password: 'YourPasswordHere'
   },
+  proddb: {
+    config: {
+      user: 'YourNameHere',
+      database: 'YourDbHere',
+      password: 'YourPasswordHere',
+      port: 5432,
+      host: 'www.yourhosthere.com',
+      ssl: true
+    }
+  },
   testdb: {
     config: {
       user: 'YourNameHere',
-      database: 'YourDBHere',
+      database: 'YourDbHere',
       password: 'YourPasswordHere',
       port: 5432,
       host: 'www.yourhosthere.com',
