+++ test/server/utils/pingUtils/emailUtilsSpec.js
@@ -5,7 +5,7 @@ var expect = chai.expect;
 var rewire = require('rewire');
 
 describe('emailUtils', function() {
-  var emailUtils = rewire('../../../server/utils/emailUtils');
+  var emailUtils = rewire('../../../../server/utils/pingUtils/emailUtils');
   emailUtils.__set__('transporter', {
     sendMail: function(mailOptions, cb) {
       cb(null, {
