+++ test/server/utils/pingUtils/smsUtilsSpec.js
@@ -5,7 +5,7 @@ var expect = chai.expect;
 var rewire = require('rewire');
 
 describe('smsUtils', function() {
-  var smsUtils = rewire('../../../server/utils/smsUtils');
+  var smsUtils = rewire('../../../../server/utils/pingUtils/smsUtils');
   smsUtils.__set__('client', {
     altsms: function(smsOptions, cb) {
       cb(null, {
